fix the efcore version error (change from 8.0.23 to 8.0.13) & change the status propery type in the sell order entity

+++ StocksAppSolution/StocksApp.Core/DTO/SellOrderDTO/SellOrderAddRequest.cs
@@ -29,7 +29,7 @@ namespace StocksApp.Core.DTO.SellOrderDTO
                 DateAndTimeOfOrder = DateAndTimeOfOrder,
                 Price = Price,
                 Quantity = Quantity,
-                Status = (int)SellOrderStatus.Pending,
+                Status = SellOrderStatus.Pending,
                 UserId = UserId
             };
         }
